# backend-innotech/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InnoTech Preinscription API",
        "description": "Documentation for the InnoTech University APIS ",
        "contact": {
            "name": "Université InnoTech",
            "email": "admissions@InnoTech-university.tn"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/formulaires/confirm-preinscription/{application}": {
            "get": {
                "tags": [
                    "Confirmation"
                ],
                "summary": "Afficher le formulaire de confirmation tout en listant les champs deja remplis dans fromualire de préinscription et les champs pour joindre les documents selon level (Licence/Mastere) mentionné dans l'appplication",
                "operationId": "b3f338cb1e88ae3d13211eeec774392a",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID de l'application",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulaire chargé avec succès"
                    },
                    "403": {
                        "description": "Déjà soumis ou non préconfirmé"
                    }
                }
            },
            "post": {
                "tags": [
                    "Confirmation"
                ],
                "summary": "Soumettre les documents de confirmation",
                "operationId": "263d6d398ad4255a300f02b25fb8ea53",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID de l'application",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fichiers requis selon le niveau",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "documents[1]": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "documents[2]": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Documents soumis avec succès"
                    },
                    "403": {
                        "description": "Formulaire déjà soumis ou non autorisé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
        "/api/formulaires/preinscription": {
            "post": {
                "tags": [
                    "Preinscription"
                ],
                "summary": "Submit a preinscription",
                "operationId": "61cb8fbff1435bca84470478c71ded3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "first_name_ar",
                                    "last_name_ar",
                                    "email",
                                    "cin"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Youssef"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Dhouib"
                                    },
                                    "first_name_ar": {
                                        "type": "string",
                                        "example": "يوسف"
                                    },
                                    "last_name_ar": {
                                        "type": "string",
                                        "example": "الذويب"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "youssef@example.com"
                                    },
                                    "cin": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "passport": {
                                        "type": "string",
                                        "example": "AB123456"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2003-05-01"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Tunisia"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Sfax"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+21698765432"
                                    },
                                    "previous_degree": {
                                        "type": "string",
                                        "example": "Baccalauréat Scientifique"
                                    },
                                    "graduation_year": {
                                        "type": "integer",
                                        "example": 2022
                                    },
                                    "how_did_you_hear": {
                                        "type": "string",
                                        "example": "Facebook"
                                    },
                                    "desired_degree_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Preinscription enregistrée avec succès"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/applications/{id}/validate": {
            "post": {
                "tags": [
                    "Preinscription"
                ],
                "summary": "Validate an application and send confirmation email",
                "operationId": "3677a8a73738257caa5e9c2bfc1d8b43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'application à valider",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Préinscription validée et email envoyé"
                    },
                    "404": {
                        "description": "Application non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
}